Refactor feedback endpoint into separate module
